implementation of ~ar, ~al, and ~alt
New aim practice commands that help players develop consistent crane control. ~ar stands for aim right, ~al for aim left, and ~alt for alternate.

These practice modes begin by spawning a single safe directly in front of the player, and progressively shifts the safe spawn in the chosen practice direction as the player hits the CFO. For alternate, your safes will alternate between left and right spawns.

You can set the vertical distance away from the toon by entering a number from -5 to 15, where -5 is close to the wall behind you, and 15 is close to the CFO. Default value is 10.